Working version

+++ Project22/myConstant.h
@@ -1,4 +1,4 @@
 #ifndef MYCONSTANT_H
 #define MYCONSTANT_H
-const float C_AccOfGrav (9.80665f);
+const double C_AccOfGrav (9.80665);
 #endif // !MYCONSTANT_H
